B/newspaper-cropper/backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyMuPDF==1.22.5
Pillow==9.5.0
